+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.6.0
+
+* Tax and benefit system evolution.
+* Impacted periods: all.
+* Impacted areas: DAF.
+* Details:
+  - Added Redevance domaniale computation
+
 ## 0.5.0
 
 * Tax and benefit system evolution.